requirements.txt
requests
datetime
customtkinter
packaging
finnhub-python